Move Code to Addon Tree
Move addon code to the addon directory and re-export it in the
app directory.

Import configuration from the consuming app's environment.js using
an initializer.
