+++ package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "broccoli-funnel": "^2.0.1",
     "broccoli-merge-trees": "^2.0.0",
+    "ember-cli-babel": "^6.0.0",
     "i18next": "^10.0.3",
     "i18next-xhr-backend": "^1.4.3"
   },
@@ -32,7 +33,6 @@
     "broccoli-asset-rev": "^2.4.5",
     "ember-ajax": "^3.0.0",
     "ember-cli": "2.13.2",
-    "ember-cli-babel": "^6.0.0",
     "ember-cli-dependency-checker": "^1.3.0",
     "ember-cli-eslint": "^3.0.0",
     "ember-cli-htmlbars": "^1.1.1",
